[JENKINS-28881] - Move Folder files. Folders belong to the model
