+++ src/main/resources/org/jenkinsci/plugins/ownership/model/folders/FolderOwnershipAction/index.jelly
