+++ src/main/resources/org/jenkinsci/plugins/ownership/model/folders/FolderOwnershipAction/manage-owners.jelly
