+++ src/main/resources/org/jenkinsci/plugins/ownership/model/folders/FolderOwnershipDescriptionSource/summary.jelly
