+++ src/main/resources/org/jenkinsci/plugins/ownership/model/folders/FolderOwnershipProperty/config.jelly
