+++ src/main/resources/org/jenkinsci/plugins/ownership/model/folders/FolderOwnershipProperty/summary.jelly
